chore: extract 44 helpers from stages.py into stage_helpers/ (AC-482)
Split loop/stages.py from 2,593 → 1,318 lines (49% reduction) by
extracting 44 private helper functions into 6 focused submodules:

- stage_helpers/freshness.py (175 lines) — freshness policy + filtering
- stage_helpers/exploration.py (342 lines) — branch generation + strategy selection
- stage_helpers/dimensions.py (180 lines) — dimension scoring + replay building
- stage_helpers/context_loaders.py (290 lines) — knowledge loading + hint/tool feedback
- stage_helpers/tournament_prep.py (142 lines) — opponent pool + holdout verification
- stage_helpers/persistence_helpers.py (295 lines) — credit assignment + curator + progress

stages.py retains 8 public stage_* functions (1,318 lines).

TDD: Added test_module_size_limits.py with 3 contract tests enforcing
file size limits and verifying the stage_helpers/ package structure.

9 files changed, 1543 insertions, 1322 deletions
