+++ autocontext/src/autocontext/loop/stages.py
@@ -10,61 +10,27 @@ from collections.abc import Callable
 from typing import TYPE_CHECKING, Any
 
 from autocontext.agents.architect import parse_dag_changes
-from autocontext.agents.feedback_loops import AnalystRating, ToolUsageTracker, format_analyst_feedback
-from autocontext.agents.hint_feedback import (
-    HintFeedback,
-    build_hint_reflection_prompt,
-    format_hint_feedback_for_coach,
-    parse_hint_feedback,
-)
-from autocontext.analytics.credit_assignment import (
-    CreditAssignmentRecord,
-    attribute_credit,
-    compute_change_vector,
-    format_attribution_for_agent,
-)
-from autocontext.harness.core.types import RoleExecution, RoleUsage
 from autocontext.harness.evaluation.dimensional import detect_dimension_regression
 from autocontext.harness.evaluation.failure_report import FailureReport
 from autocontext.harness.evaluation.runner import EvaluationRunner
 from autocontext.harness.evaluation.scenario_evaluator import ScenarioEvaluator
-from autocontext.harness.evaluation.self_play import (
-    SelfPlayConfig,
-    build_opponent_pool,
-    load_self_play_pool,
-)
 from autocontext.harness.evaluation.types import EvaluationLimits as HarnessLimits
-from autocontext.harness.evaluation.types import EvaluationResult, EvaluationSummary
-from autocontext.harness.pipeline.holdout import HoldoutPolicy, HoldoutResult, HoldoutVerifier
+from autocontext.harness.evaluation.types import EvaluationResult
+from autocontext.harness.pipeline.holdout import HoldoutResult
 from autocontext.harness.pipeline.trend_gate import TrendAwareGate
 from autocontext.harness.pipeline.validity_gate import ValidityGate
-from autocontext.knowledge.dead_end_manager import DeadEndEntry, consolidate_dead_ends
-from autocontext.knowledge.evidence_freshness import (
-    EvidenceFreshness,
-    FreshnessPolicy,
-    apply_freshness_decay,
-    detect_stale_context,
-)
 from autocontext.knowledge.fresh_start import execute_fresh_start
 from autocontext.knowledge.harness_quality import compute_harness_quality
-from autocontext.knowledge.hint_volume import HintManager, HintVolumePolicy
-from autocontext.knowledge.progress import build_progress_snapshot
+from autocontext.knowledge.hint_volume import HintManager
 from autocontext.knowledge.protocol import parse_research_protocol, validate_tuning_overrides
 from autocontext.knowledge.rapid_gate import rapid_gate, should_transition_to_linear
 from autocontext.knowledge.stagnation import StagnationDetector
 from autocontext.knowledge.tuning import TuningConfig, parse_tuning_proposal
 from autocontext.loop.cost_control import CostPolicy, evaluate_cost_effectiveness
 from autocontext.loop.exploration import (
-    BasinCandidate,
-    BranchRecord,
-    DivergentCompetitorConfig,
-    MultiBasinConfig,
     NoveltyConfig,
     apply_novelty_bonus,
     compute_novelty_score,
-    generate_basin_candidates,
-    should_spawn_divergent,
-    should_trigger_multi_basin,
 )
 from autocontext.loop.stage_types import GenerationContext
 from autocontext.loop.tournament_helpers import (
@@ -77,7 +43,6 @@ from autocontext.notebook.context_provider import NotebookContextProvider
 from autocontext.notebook.types import SessionNotebook
 from autocontext.prompts.templates import build_prompt_bundle
 from autocontext.providers.base import CompletionResult, LLMProvider
-from autocontext.scenarios.families import detect_family
 from autocontext.storage.artifacts import EMPTY_PLAYBOOK_SENTINEL
 
 if TYPE_CHECKING:
@@ -85,172 +50,59 @@ if TYPE_CHECKING:
     from autocontext.agents.llm_client import LanguageModelClient
     from autocontext.agents.orchestrator import AgentOrchestrator
     from autocontext.agents.skeptic import SkepticAgent
-    from autocontext.agents.types import AgentOutputs
     from autocontext.execution.supervisor import ExecutionSupervisor
     from autocontext.harness.pipeline.gate import BackpressureGate
     from autocontext.knowledge.trajectory import ScoreTrajectoryBuilder
     from autocontext.loop.events import EventStreamEmitter
     from autocontext.storage import ArtifactStore, SQLiteStore
 
+from autocontext.loop.stage_helpers.context_loaders import (
+    _apply_hint_feedback_to_manager,
+    _collect_hint_feedback,
+    _hint_volume_policy,
+    _load_analyst_feedback_section,
+    _load_architect_tool_usage_report,
+    _load_credit_attribution_section,
+    _load_hint_feedback_section,
+    _load_validity_harness_loader,
+    _normalize_tool_names,
+    _update_tool_usage_feedback,
+)
+from autocontext.loop.stage_helpers.dimensions import (
+    _build_dimension_summary_payload,
+    _build_match_replay_json,
+    _build_replay_envelope_payload,
+    _build_self_play_summary_payload,
+    _load_previous_best_dimensions,
+)
+from autocontext.loop.stage_helpers.exploration import (
+    _load_recent_numeric_strategies,
+    _select_exploration_strategy,
+)
+from autocontext.loop.stage_helpers.freshness import (
+    _filter_notebook_by_freshness,
+    _load_fresh_hint_context,
+    _load_fresh_skill_context,
+)
+from autocontext.loop.stage_helpers.persistence_helpers import (
+    _apply_tuning_to_settings,
+    _build_credit_assignment_record,
+    _maybe_rate_analyst_output,
+    _persist_progress_snapshot,
+    _persist_skill_note,
+    _revise_strategy_for_validity_failure,
+    _run_curator_consolidation,
+)
+from autocontext.loop.stage_helpers.tournament_prep import (
+    _build_empty_tournament,
+    _build_live_opponent_pool,
+    _build_skeptic_review_section,
+    _run_holdout_verification,
+)
+
 logger = logging.getLogger(__name__)
 
 _NOTEBOOK_CONTEXT_PROVIDER = NotebookContextProvider()
-
-
-def _freshness_policy(ctx: GenerationContext) -> FreshnessPolicy:
-    return FreshnessPolicy(
-        max_age_gens=ctx.settings.evidence_freshness_max_age_gens,
-        min_confidence=ctx.settings.evidence_freshness_min_confidence,
-        min_support=ctx.settings.evidence_freshness_min_support,
-    )
-
-
-def _format_freshness_warning_block(label: str, warnings: list[str]) -> str:
-    if not warnings:
-        return ""
-    return f"{label} freshness warnings:\n" + "\n".join(f"- {warning}" for warning in warnings)
-
-
-def _load_fresh_skill_context(
-    ctx: GenerationContext,
-    *,
-    artifacts: ArtifactStore,
-) -> tuple[str, str]:
-    lessons = artifacts.lesson_store.read_lessons(ctx.scenario_name)
-    if not lessons:
-        return artifacts.read_skills(ctx.scenario_name), ""
-
-    records: list[tuple[str, EvidenceFreshness]] = []
-    for lesson in lessons:
-        records.append((
-            lesson.text.strip(),
-            EvidenceFreshness(
-                item_id=lesson.id or lesson.text.strip(),
-                support_count=1,
-                last_validated_gen=max(lesson.meta.last_validated_gen, 0),
-                confidence=max(0.0, min(1.0, lesson.meta.best_score)),
-                created_at_gen=max(lesson.meta.generation, 0),
-            ),
-        ))
-
-    items = [item for _, item in records]
-    active, _ = apply_freshness_decay(items, ctx.generation, _freshness_policy(ctx))
-    active_ids = {item.item_id for item in active}
-    active_text = "\n".join(
-        text for text, item in records if item.item_id in active_ids
-    ).strip()
-    warnings = detect_stale_context(items, ctx.generation, _freshness_policy(ctx))
-    return active_text, _format_freshness_warning_block("Lesson", warnings)
-
-
-def _load_fresh_hint_context(
-    ctx: GenerationContext,
-    *,
-    artifacts: ArtifactStore,
-) -> tuple[str, str]:
-    if not ctx.settings.hint_volume_enabled:
-        return ctx.coach_competitor_hints, ""
-
-    raw_manager = artifacts.read_hint_manager(
-        ctx.scenario_name,
-        policy=HintVolumePolicy(
-            max_hints=ctx.settings.hint_volume_max_hints,
-            archive_rotated=ctx.settings.hint_volume_archive_rotated,
-        ),
-    )
-    manager = raw_manager if isinstance(raw_manager, HintManager) else HintManager(HintVolumePolicy())
-    ranked_hints = manager.active_hints()
-    if not ranked_hints:
-        return "", ""
-
-    records: list[tuple[str, EvidenceFreshness]] = []
-    for hint in ranked_hints:
-        records.append((
-            hint.text,
-            EvidenceFreshness(
-                item_id=hint.text,
-                support_count=1,
-                last_validated_gen=max(hint.generation_added, 0),
-                confidence=max(0.0, min(1.0, hint.impact_score)),
-                created_at_gen=max(hint.generation_added, 0),
-            ),
-        ))
-
-    items = [item for _, item in records]
-    active, _ = apply_freshness_decay(items, ctx.generation, _freshness_policy(ctx))
-    active_ids = {item.item_id for item in active}
-    fresh_hints = "\n".join(
-        f"- {text}" for text, item in records if item.item_id in active_ids
-    ).strip()
-    warnings = detect_stale_context(items, ctx.generation, _freshness_policy(ctx))
-    return fresh_hints, _format_freshness_warning_block("Hint", warnings)
-
-
-def _filter_notebook_by_freshness(
-    ctx: GenerationContext,
-    notebook: SessionNotebook,
-) -> tuple[SessionNotebook, str]:
-    last_validated_gen = notebook.best_generation if notebook.best_generation is not None else ctx.generation
-    confidence = notebook.best_score if notebook.best_score is not None else 1.0
-    fields = [
-        "current_objective",
-        "current_hypotheses",
-        "unresolved_questions",
-        "operator_observations",
-        "follow_ups",
-    ]
-    records: list[tuple[str, EvidenceFreshness]] = []
-    for field_name in fields:
-        value = getattr(notebook, field_name)
-        if not value:
-            continue
-        records.append((
-            field_name,
-            EvidenceFreshness(
-                item_id=f"notebook:{field_name}",
-                support_count=1,
-                last_validated_gen=max(last_validated_gen, 0),
-                confidence=max(0.0, min(1.0, confidence)),
-                created_at_gen=max(last_validated_gen, 0),
-            ),
-        ))
-
-    if not records:
-        return notebook, ""
-
-    items = [item for _, item in records]
-    active, _ = apply_freshness_decay(items, ctx.generation, _freshness_policy(ctx))
-    active_ids = {item.item_id for item in active}
-    filtered = dataclasses.replace(
-        notebook,
-        current_objective=(
-            notebook.current_objective
-            if "notebook:current_objective" in active_ids
-            else ""
-        ),
-        current_hypotheses=(
-            notebook.current_hypotheses
-            if "notebook:current_hypotheses" in active_ids
-            else []
-        ),
-        unresolved_questions=(
-            notebook.unresolved_questions
-            if "notebook:unresolved_questions" in active_ids
-            else []
-        ),
-        operator_observations=(
-            notebook.operator_observations
-            if "notebook:operator_observations" in active_ids
-            else []
-        ),
-        follow_ups=(
-            notebook.follow_ups
-            if "notebook:follow_ups" in active_ids
-            else []
-        ),
-    )
-    warnings = detect_stale_context(items, ctx.generation, _freshness_policy(ctx))
-    return filtered, _format_freshness_warning_block("Notebook", warnings)
 
 
 class _ClientAsProvider(LLMProvider):
@@ -278,624 +130,6 @@ class _ClientAsProvider(LLMProvider):
 
     def default_model(self) -> str:
         return self._model
-
-
-def _load_validity_harness_loader(
-    ctx: GenerationContext,
-    *,
-    artifacts: ArtifactStore,
-) -> Any | None:
-    """Load harness validators for two-tier validity checks when enabled."""
-    if not ctx.settings.harness_validators_enabled:
-        return None
-
-    from autocontext.execution.harness_loader import HarnessLoader
-
-    harness_dir = artifacts.harness_dir(ctx.scenario_name)
-    if not harness_dir.exists():
-        return None
-
-    loader = HarnessLoader(
-        harness_dir,
-        timeout_seconds=ctx.settings.harness_timeout_seconds,
-    )
-    loader.load()
-    return loader
-
-
-def _build_empty_tournament(ctx: GenerationContext) -> EvaluationSummary:
-    """Create a zero-match summary for rollback paths that skip execution."""
-    return EvaluationSummary(
-        mean_score=0.0,
-        best_score=0.0,
-        wins=0,
-        losses=0,
-        elo_after=ctx.challenger_elo,
-        results=[],
-        scoring_backend=ctx.settings.scoring_backend,
-        uncertainty_after=ctx.challenger_uncertainty,
-    )
-
-
-def _build_live_opponent_pool(
-    ctx: GenerationContext,
-    *,
-    sqlite: SQLiteStore,
-) -> tuple[Any, list[dict[str, Any]], int]:
-    """Build the same opponent schedule used by the live tournament path."""
-    settings = ctx.settings
-    self_play_config = SelfPlayConfig(
-        enabled=settings.self_play_enabled,
-        pool_size=settings.self_play_pool_size,
-        weight=settings.self_play_weight,
-    )
-    self_play_pool = load_self_play_pool(
-        sqlite.get_self_play_strategy_history(ctx.run_id) if settings.self_play_enabled else [],
-        self_play_config,
-        current_generation=ctx.generation,
-    )
-    opponent_pool = build_opponent_pool(
-        [{"source": "baseline"}],
-        self_play_pool,
-        trials=settings.matches_per_generation,
-    )
-    planned_self_play_matches = sum(
-        1
-        for entry in opponent_pool
-        if isinstance(entry, dict) and entry.get("source") == "self_play"
-    )
-    return self_play_pool, opponent_pool, planned_self_play_matches
-
-
-def _load_recent_numeric_strategies(
-    sqlite: SQLiteStore,
-    *,
-    run_id: str,
-    window: int,
-) -> list[dict[str, Any]]:
-    """Load recent persisted competitor strategies for novelty comparison."""
-    try:
-        history = sqlite.get_strategy_score_history(run_id)
-    except Exception:
-        logger.debug("failed to load strategy history for novelty", exc_info=True)
-        return []
-
-    recent: list[dict[str, Any]] = []
-    for row in history[-window:]:
-        if not isinstance(row, dict):
-            continue
-        raw_content = row.get("content")
-        if not isinstance(raw_content, str) or not raw_content.strip():
-            continue
-        try:
-            parsed = json.loads(raw_content)
-        except json.JSONDecodeError:
-            continue
-        if isinstance(parsed, dict):
-            recent.append(parsed)
-    return recent
-
-
-def _replace_prompt_section(
-    prompt: str,
-    *,
-    label: str,
-    old_value: str,
-    new_value: str,
-    anchor_label: str | None = None,
-) -> str:
-    old_block = f"{label}:\n{old_value}\n\n" if old_value else ""
-    new_block = f"{label}:\n{new_value}\n\n" if new_value else ""
-    if old_block and old_block in prompt:
-        return prompt.replace(old_block, new_block, 1)
-    if old_block:
-        return prompt
-    if not new_block:
-        return prompt
-    if anchor_label:
-        anchor = f"{anchor_label}:\n"
-        index = prompt.find(anchor)
-        if index >= 0:
-            block_end = prompt.find("\n\n", index)
-            if block_end >= 0:
-                insert_at = block_end + 2
-                return prompt[:insert_at] + new_block + prompt[insert_at:]
-    return prompt
-
-
-def _build_branch_competitor_prompt(
-    ctx: GenerationContext,
-    *,
-    playbook: str,
-    lessons: str,
-    note: str = "",
-) -> str:
-    if ctx.prompts is None:
-        raise RuntimeError("stage_knowledge_setup must run first")
-
-    prompt = _replace_prompt_section(
-        ctx.prompts.competitor,
-        label="Current playbook",
-        old_value=ctx.base_playbook,
-        new_value=playbook,
-    )
-    prompt = _replace_prompt_section(
-        prompt,
-        label="Operational lessons (from prior generations)",
-        old_value=ctx.base_lessons,
-        new_value=lessons,
-        anchor_label="Current playbook",
-    )
-    if note:
-        prompt += f"\n\nExploration branch note:\n{note}"
-    return prompt
-
-
-def _generate_branch_strategy(
-    ctx: GenerationContext,
-    *,
-    orchestrator: AgentOrchestrator,
-    prompt: str,
-    temperature: float,
-) -> tuple[dict[str, Any], RoleExecution, RoleExecution]:
-    """Run competitor + translator for a single exploration branch."""
-    if ctx.prompts is None:
-        raise RuntimeError("stage_knowledge_setup must run first")
-
-    competitor_prompt = prompt
-    if ctx.settings.code_strategies_enabled:
-        from autocontext.prompts.templates import code_strategy_competitor_suffix
-
-        competitor_prompt += code_strategy_competitor_suffix(ctx.strategy_interface)
-
-    with orchestrator._use_role_runtime(  # noqa: SLF001 - stage needs routed role runtime
-        "competitor",
-        orchestrator.competitor,
-        generation=ctx.generation,
-        scenario_name=ctx.scenario_name,
-    ):
-        raw_text, competitor_exec = orchestrator.competitor.run(
-            competitor_prompt,
-            tool_context=ctx.tool_context,
-            temperature=temperature,
-        )
-    with orchestrator._use_role_runtime(  # noqa: SLF001 - stage needs routed role runtime
-        "translator",
-        orchestrator.translator,
-        generation=ctx.generation,
-        scenario_name=ctx.scenario_name,
-    ):
-        if ctx.settings.code_strategies_enabled:
-            strategy, translator_exec = orchestrator.translator.translate_code(raw_text)
-        else:
-            strategy, translator_exec = orchestrator.translator.translate(raw_text, ctx.strategy_interface)
-    return strategy, competitor_exec, translator_exec
-
-
-def _select_exploration_strategy(
-    ctx: GenerationContext,
-    *,
-    outputs: AgentOutputs,
-    orchestrator: AgentOrchestrator,
-    supervisor: ExecutionSupervisor | None,
-    sqlite: SQLiteStore,
-    events: EventStreamEmitter | None,
-) -> tuple[dict[str, Any], dict[str, Any]]:
-    """Optionally explore multiple competitor basins and return the selected strategy."""
-    settings = ctx.settings
-    if supervisor is None:
-        return outputs.strategy, {}
-
-    multi_basin_config = MultiBasinConfig(
-        enabled=settings.multi_basin_enabled,
-        trigger_rollbacks=settings.multi_basin_trigger_rollbacks,
-        candidates=settings.multi_basin_candidates,
-        periodic_every_n=settings.multi_basin_periodic_every_n,
-    )
-    divergent_config = DivergentCompetitorConfig(
-        enabled=settings.divergent_competitor_enabled,
-        rollback_threshold=settings.divergent_rollback_threshold,
-        temperature=settings.divergent_temperature,
-    )
-    multi_basin_triggered = should_trigger_multi_basin(
-        ctx.gate_decision_history,
-        ctx.generation,
-        multi_basin_config,
-    )
-    divergent_triggered = should_spawn_divergent(ctx.gate_decision_history, divergent_config)
-
-    if not multi_basin_triggered and not divergent_triggered:
-        return outputs.strategy, {}
-
-    branch_specs: list[BasinCandidate] = []
-    if multi_basin_triggered:
-        branch_specs = generate_basin_candidates(
-            ctx.base_playbook,
-            ctx.base_lessons,
-            multi_basin_config,
-        )
-    else:
-        branch_specs = [
-            BasinCandidate(
-                branch_type="conservative",
-                playbook=ctx.base_playbook,
-                lessons=ctx.base_lessons,
-                temperature=0.2,
-            ),
-            BasinCandidate(
-                branch_type="divergent",
-                playbook="",
-                lessons=ctx.base_lessons,
-                temperature=divergent_config.temperature,
-                metadata={"note": "Fresh start with lessons only"},
-            ),
-        ]
-
-    candidate_entries: list[dict[str, Any]] = [{
-        "branch_type": "conservative",
-        "strategy": outputs.strategy,
-        "temperature": 0.2,
-        "metadata": {"source": "base_generation"},
-    }]
-    seen_strategies = {json.dumps(outputs.strategy, sort_keys=True)}
-
-    if events is not None:
-        events.emit("exploration_started", {
-            "run_id": ctx.run_id,
-            "generation": ctx.generation,
-            "multi_basin_triggered": multi_basin_triggered,
-            "divergent_triggered": divergent_triggered,
-            "gate_history": ctx.gate_decision_history,
-        })
-
-    for branch in branch_specs:
-        if branch.branch_type == "conservative":
-            continue
-        branch_temperature = (
-            divergent_config.temperature
-            if branch.branch_type == "divergent"
-            else branch.temperature
-        )
-        branch_prompt = _build_branch_competitor_prompt(
-            ctx,
-            playbook=branch.playbook,
-            lessons=branch.lessons,
-            note=str(branch.metadata.get("note", "")),
-        )
-        try:
-            strategy, _, _ = _generate_branch_strategy(
-                ctx,
-                orchestrator=orchestrator,
-                prompt=branch_prompt,
-                temperature=branch_temperature,
-            )
-        except Exception:
-            logger.debug("failed to generate %s exploration branch", branch.branch_type, exc_info=True)
-            continue
-
-        serialized = json.dumps(strategy, sort_keys=True)
-        if serialized in seen_strategies:
-            continue
-        if "__code__" not in strategy:
-            state = ctx.scenario.initial_state(seed=settings.seed_base + ctx.generation)
-            valid, _reason = ctx.scenario.validate_actions(state, "challenger", strategy)
-            if not valid:
-                continue
-        seen_strategies.add(serialized)
-        candidate_entries.append({
-            "branch_type": branch.branch_type,
-            "strategy": strategy,
-            "temperature": branch_temperature,
-            "metadata": dict(branch.metadata),
-        })
-
-    if len(candidate_entries) == 1:
-        return outputs.strategy, {}
-
-    _self_play_pool, opponent_pool, planned_self_play_matches = _build_live_opponent_pool(ctx, sqlite=sqlite)
-    evaluator = ScenarioEvaluator(ctx.scenario, supervisor)
-    runner = EvaluationRunner(evaluator, scoring_backend=settings.scoring_backend)
-    selection_results: list[dict[str, Any]] = []
-
-    for candidate in candidate_entries:
-        tournament = runner.run(
-            candidate=candidate["strategy"],
-            seed_base=settings.seed_base + (ctx.generation * 100),
-            trials=settings.matches_per_generation,
-            limits=HarnessLimits(),
-            challenger_elo=ctx.challenger_elo,
-            challenger_uncertainty=ctx.challenger_uncertainty,
-            opponent_pool=opponent_pool,
-        )
-        selection_results.append({
-            "branch_type": candidate["branch_type"],
-            "best_score": tournament.best_score,
-            "mean_score": tournament.mean_score,
-            "strategy": candidate["strategy"],
-            "temperature": candidate["temperature"],
-            "metadata": dict(candidate.get("metadata", {})),
-        })
-
-    selected = max(
-        selection_results,
-        key=lambda item: (float(item["best_score"]), float(item["mean_score"])),
-    )
-    branch_record = BranchRecord(
-        generation=ctx.generation,
-        branch_type=str(selected["branch_type"]),
-        score=float(selected["best_score"]),
-        advanced=False,
-        metadata={
-            "selection_mean_score": float(selected["mean_score"]),
-            "selection_match_count": settings.matches_per_generation,
-            "self_play_matches_planned": planned_self_play_matches,
-            "multi_basin_triggered": multi_basin_triggered,
-            "divergent_triggered": divergent_triggered,
-        },
-    )
-    metadata = {
-        "selected_branch": branch_record.to_dict(),
-        "candidates": [
-            {
-                "branch_type": str(item["branch_type"]),
-                "best_score": float(item["best_score"]),
-                "mean_score": float(item["mean_score"]),
-                "temperature": float(item["temperature"]),
-                "metadata": dict(item["metadata"]),
-            }
-            for item in selection_results
-        ],
-    }
-    if events is not None:
-        events.emit("exploration_selected", {
-            "run_id": ctx.run_id,
-            "generation": ctx.generation,
-            **metadata,
-        })
-    return dict(selected["strategy"]), metadata
-
-
-def _load_previous_best_dimensions(
-    sqlite: SQLiteStore,
-    run_id: str,
-) -> dict[str, float]:
-    """Read the latest persisted generation dimensions for regression comparison."""
-    try:
-        rows = sqlite.get_generation_trajectory(run_id)
-    except Exception:
-        logger.debug("failed to load previous dimension summary", exc_info=True)
-        return {}
-    if not isinstance(rows, list) or not rows:
-        return {}
-    latest = rows[-1]
-    if not isinstance(latest, dict):
-        return {}
-    summary = latest.get("dimension_summary")
-    if not isinstance(summary, dict):
-        return {}
-    raw_best = summary.get("best_dimensions")
-    if not isinstance(raw_best, dict):
-        return {}
-    return {
-        name: float(value)
-        for name, value in raw_best.items()
-        if isinstance(name, str) and isinstance(value, (int, float))
-    }
-
-
-def _coerce_dimension_score_map(raw_value: Any) -> dict[str, float]:
-    """Return a JSON-safe dimension score mapping."""
-    if not isinstance(raw_value, dict):
-        return {}
-    return {
-        name: round(float(value), 6)
-        for name, value in raw_value.items()
-        if isinstance(name, str) and isinstance(value, (int, float))
-    }
-
-
-def _coerce_dimension_specs(raw_value: Any) -> list[dict[str, Any]]:
-    """Return JSON-safe dimension specs."""
-    if not isinstance(raw_value, list):
-        return []
-    specs: list[dict[str, Any]] = []
-    for item in raw_value:
-        if not isinstance(item, dict):
-            continue
-        clean: dict[str, Any] = {
-            key: value
-            for key, value in item.items()
-            if isinstance(key, str)
-            and (value is None or isinstance(value, (str, int, float, bool)))
-        }
-        if clean:
-            specs.append(clean)
-    return specs
-
-
-def _coerce_dimension_regressions(raw_value: Any) -> list[dict[str, Any]]:
-    """Return JSON-safe dimension regression payloads."""
-    if not isinstance(raw_value, list):
-        return []
-    regressions: list[dict[str, Any]] = []
-    for item in raw_value:
-        if not isinstance(item, dict):
-            continue
-        dimension = item.get("dimension")
-        previous = item.get("previous")
-        current = item.get("current")
-        delta = item.get("delta")
-        if not isinstance(dimension, str):
-            continue
-        if not isinstance(previous, (int, float)):
-            continue
-        if not isinstance(current, (int, float)):
-            continue
-        if not isinstance(delta, (int, float)):
-            continue
-        regressions.append({
-            "dimension": dimension,
-            "previous": round(float(previous), 6),
-            "current": round(float(current), 6),
-            "delta": round(float(delta), 6),
-        })
-    return regressions
-
-
-def _build_dimension_summary_payload(tournament: EvaluationSummary) -> dict[str, Any] | None:
-    """Extract a JSON-safe dimensional summary from a tournament."""
-    dimension_means = _coerce_dimension_score_map(getattr(tournament, "dimension_means", {}))
-    best_dimensions = _coerce_dimension_score_map(getattr(tournament, "best_dimensions", {}))
-    dimension_specs = _coerce_dimension_specs(getattr(tournament, "dimension_specs", []))
-    dimension_regressions = _coerce_dimension_regressions(
-        getattr(tournament, "dimension_regressions", []),
-    )
-    if not any((dimension_means, best_dimensions, dimension_specs, dimension_regressions)):
-        return None
-    return {
-        "dimension_means": dimension_means,
-        "best_dimensions": best_dimensions,
-        "dimension_specs": dimension_specs,
-        "dimension_regressions": dimension_regressions,
-    }
-
-
-def _build_self_play_summary_payload(tournament: EvaluationSummary) -> dict[str, Any] | None:
-    """Extract a JSON-safe self-play summary from a tournament."""
-    raw_value = getattr(tournament, "self_play_summary", {})
-    if not isinstance(raw_value, dict):
-        return None
-    clean: dict[str, Any] = {}
-    for key, value in raw_value.items():
-        if not isinstance(key, str):
-            continue
-        if isinstance(value, (bool, str)):
-            clean[key] = value
-            continue
-        if isinstance(value, int):
-            clean[key] = value
-            continue
-        if isinstance(value, float):
-            clean[key] = round(value, 6)
-    return clean or None
-
-
-def _json_dumps_if_serializable(value: Any) -> str | None:
-    try:
-        return json.dumps(value)
-    except (TypeError, ValueError):
-        return None
-
-
-def _build_replay_envelope_payload(execution_output: Any) -> dict[str, Any]:
-    replay = getattr(execution_output, "replay", None)
-    model_dump = getattr(replay, "model_dump", None)
-    if not callable(model_dump):
-        return {}
-    try:
-        payload = model_dump()
-    except Exception:
-        logger.debug("loop.stages: caught Exception", exc_info=True)
-        return {}
-    if not isinstance(payload, dict):
-        return {}
-    if _json_dumps_if_serializable(payload) is None:
-        return {}
-    return payload
-
-
-def _build_match_replay_json(execution_output: Any) -> str:
-    result = getattr(execution_output, "result", None)
-    replay = getattr(result, "replay", None)
-    if replay:
-        serialized = _json_dumps_if_serializable(replay)
-        if serialized is not None:
-            return serialized
-
-    replay_payload = _build_replay_envelope_payload(execution_output)
-    timeline = replay_payload.get("timeline")
-    if timeline:
-        serialized = _json_dumps_if_serializable(timeline)
-        if serialized is not None:
-            return serialized
-    return ""
-
-
-def _build_skeptic_review_section(ctx: GenerationContext) -> str:
-    """Render skeptic findings into curator-readable context."""
-    review = ctx.skeptic_review
-    if review is None:
-        return ""
-    concerns = review.concerns or ["No concrete concerns captured."]
-    concerns_block = "\n".join(f"- {concern}" for concern in concerns)
-    return (
-        "SKEPTIC REVIEW:\n"
-        f"Risk level: {review.risk_level}\n"
-        f"Recommendation: {review.recommendation}\n"
-        f"Confidence: {review.confidence}/10\n"
-        "Concerns:\n"
-        f"{concerns_block}\n"
-    )
-
-
-def _resolve_holdout_policy(ctx: GenerationContext) -> HoldoutPolicy:
-    """Build the effective holdout policy, including scenario-family overrides."""
-    family = detect_family(ctx.scenario)
-    family_marker = family.scenario_type_marker if family is not None else ""
-    policy = HoldoutPolicy(
-        holdout_seeds=ctx.settings.holdout_seeds,
-        min_holdout_score=ctx.settings.holdout_min_score,
-        max_generalization_gap=ctx.settings.holdout_max_regression_gap,
-        seed_offset=ctx.settings.holdout_seed_offset,
-        enabled=ctx.settings.holdout_enabled,
-        metadata={"family": family_marker} if family_marker else {},
-    )
-    if family is None:
-        return policy
-
-    override = (
-        ctx.settings.holdout_family_policies.get(family.scenario_type_marker)
-        or ctx.settings.holdout_family_policies.get(family.name)
-    )
-    if not isinstance(override, dict):
-        return policy
-
-    merged = policy.to_dict()
-    merged.update(override)
-    metadata = dict(policy.metadata)
-    override_metadata = override.get("metadata")
-    if isinstance(override_metadata, dict):
-        metadata.update(override_metadata)
-    if family_marker:
-        metadata.setdefault("family", family_marker)
-    merged["metadata"] = metadata
-    return HoldoutPolicy.from_dict(merged)
-
-
-def _run_holdout_verification(
-    ctx: GenerationContext,
-    *,
-    supervisor: ExecutionSupervisor,
-    strategy: dict[str, Any],
-    in_sample_score: float,
-    limits: HarnessLimits,
-) -> HoldoutResult | None:
-    """Verify an advancing candidate on holdout seeds when enabled."""
-    policy = _resolve_holdout_policy(ctx)
-    if not policy.enabled:
-        return None
-
-    evaluator = ScenarioEvaluator(ctx.scenario, supervisor)
-
-    def _evaluate(candidate: dict[str, Any], seed: int) -> float:
-        return evaluator.evaluate(candidate, seed, limits).score
-
-    verifier = HoldoutVerifier(policy=policy, evaluate_fn=_evaluate)
-    result = verifier.verify(strategy=strategy, in_sample_score=in_sample_score)
-    metadata = dict(result.metadata)
-    metadata["policy"] = policy.to_dict()
-    result.metadata = metadata
-    return result
 
 
 def stage_policy_refinement(
@@ -981,398 +215,6 @@ def stage_policy_refinement(
         })
 
     return ctx
-
-
-def _revise_strategy_for_validity_failure(
-    ctx: GenerationContext,
-    *,
-    current_strategy: dict[str, Any],
-    errors: list[str],
-    retry_attempt: int,
-    agents: AgentOrchestrator | None,
-) -> dict[str, Any] | None:
-    """Ask the competitor to fix an invalid strategy before running matches."""
-    if agents is None or ctx.prompts is None:
-        return None
-
-    is_code_strategy = "__code__" in current_strategy
-    retry_prompt = (
-        ctx.prompts.competitor
-        + f"\n\n--- VALIDITY RETRY ATTEMPT {retry_attempt} ---\n"
-        + "Your previous strategy failed pre-tournament validation.\n"
-        + "Validation errors:\n"
-        + "\n".join(f"- {error}" for error in errors)
-        + "\n"
-    )
-    if is_code_strategy:
-        retry_prompt += "Adjust your code so it satisfies the harness and scenario contracts.\n"
-        if ctx.settings.code_strategies_enabled:
-            from autocontext.prompts.templates import code_strategy_competitor_suffix
-
-            retry_prompt += code_strategy_competitor_suffix(ctx.strategy_interface)
-    else:
-        retry_prompt += (
-            f"Previous strategy: {json.dumps(current_strategy, sort_keys=True)}\n"
-            "Return a revised valid strategy. Do not repeat the same invalid approach.\n"
-        )
-
-    try:
-        raw_text, _ = agents.competitor.run(retry_prompt, tool_context=ctx.tool_context)
-        if is_code_strategy:
-            revised_strategy, _ = agents.translator.translate_code(raw_text)
-        else:
-            revised_strategy, _ = agents.translator.translate(raw_text, ctx.strategy_interface)
-        return revised_strategy
-    except Exception:
-        logger.debug("validity retry competitor re-invocation failed", exc_info=True)
-        return None
-
-
-def _apply_tuning_to_settings(
-    ctx: GenerationContext,
-    parameters: dict[str, float | int],
-) -> None:
-    """Apply validated tuning parameters to ctx.settings (Pydantic model copy)."""
-    if not parameters:
-        return
-    update: dict[str, Any] = {}
-    for key, value in parameters.items():
-        if hasattr(ctx.settings, key):
-            update[key] = value
-    if update:
-        ctx.settings = ctx.settings.model_copy(update=update)
-
-
-def _load_analyst_feedback_section(
-    ctx: GenerationContext,
-    *,
-    artifacts: ArtifactStore,
-) -> str:
-    """Read the latest curator rating for injection into the next analyst prompt."""
-    raw_rating = artifacts.read_latest_analyst_rating(ctx.scenario_name, ctx.generation)
-    if not isinstance(raw_rating, AnalystRating):
-        return ""
-    return format_analyst_feedback(raw_rating)
-
-
-def _load_architect_tool_usage_report(
-    ctx: GenerationContext,
-    *,
-    artifacts: ArtifactStore,
-) -> str:
-    """Read the architect-facing report on which tools the competitor actually uses."""
-    if ctx.generation <= 1:
-        return ""
-    report = artifacts.read_tool_usage_report(
-        ctx.scenario_name,
-        current_generation=ctx.generation - 1,
-    )
-    return report if isinstance(report, str) else ""
-
-
-def _load_hint_feedback_section(
-    ctx: GenerationContext,
-    *,
-    artifacts: ArtifactStore,
-) -> str:
-    """Read the latest competitor feedback so the next coach prompt can use it."""
-    if ctx.generation <= 1:
-        return ""
-    raw_feedback = artifacts.read_latest_hint_feedback(ctx.scenario_name, ctx.generation)
-    if not isinstance(raw_feedback, HintFeedback):
-        return ""
-    return format_hint_feedback_for_coach(raw_feedback)
-
-
-def _load_credit_attribution_section(
-    ctx: GenerationContext,
-    *,
-    artifacts: ArtifactStore,
-    role: str,
-) -> str:
-    """Read the latest attribution record and format it for a specific agent role."""
-    if ctx.generation <= 1:
-        return ""
-    raw_record = artifacts.read_latest_credit_assignment(
-        ctx.scenario_name,
-        run_id=ctx.run_id,
-        current_gen=ctx.generation,
-    )
-    if not isinstance(raw_record, CreditAssignmentRecord):
-        return ""
-    return format_attribution_for_agent(raw_record.attribution, role)
-
-
-def _normalize_tool_names(raw: object) -> list[str]:
-    """Normalize tool names from persisted lists or created-tool markers."""
-    if not isinstance(raw, list):
-        return []
-    normalized: list[str] = []
-    for item in raw:
-        if not isinstance(item, str):
-            continue
-        value = item.strip()
-        if not value:
-            continue
-        if value.endswith(" (updated)"):
-            value = value[: -len(" (updated)")]
-        if value.endswith(".py"):
-            value = value[:-3]
-        if value and value not in normalized:
-            normalized.append(value)
-    return normalized
-
-
-def _current_tool_names(ctx: GenerationContext, *, artifacts: ArtifactStore) -> list[str]:
-    """Return the persisted post-generation tool set with a safe fallback for tests."""
-    if hasattr(artifacts, "list_tool_names"):
-        raw_names = artifacts.list_tool_names(ctx.scenario_name)
-        normalized = _normalize_tool_names(raw_names)
-        if normalized:
-            return normalized
-    merged = [*ctx.base_tool_names, *_normalize_tool_names(ctx.created_tools)]
-    deduped: list[str] = []
-    for name in merged:
-        if name not in deduped:
-            deduped.append(name)
-    return deduped
-
-
-def _build_credit_assignment_record(
-    ctx: GenerationContext,
-    *,
-    artifacts: ArtifactStore,
-) -> CreditAssignmentRecord | None:
-    """Compute a durable attribution record from the persisted generation state."""
-    outputs = ctx.outputs
-    if outputs is None:
-        return None
-
-    score_delta = ctx.gate_delta
-    previous_state = {
-        "playbook": ctx.base_playbook,
-        "tools": ctx.base_tool_names,
-        "hints": ctx.applied_competitor_hints,
-        "analysis": ctx.base_analysis,
-    }
-    current_state = {
-        "playbook": outputs.coach_playbook if ctx.gate_decision == "advance" else ctx.base_playbook,
-        "tools": _current_tool_names(ctx, artifacts=artifacts),
-        "hints": ctx.coach_competitor_hints,
-        "analysis": outputs.analysis_markdown,
-    }
-    vector = compute_change_vector(
-        generation=ctx.generation,
-        score_delta=score_delta,
-        previous_state=previous_state,
-        current_state=current_state,
-    )
-    attribution = attribute_credit(vector)
-    return CreditAssignmentRecord(
-        run_id=ctx.run_id,
-        generation=ctx.generation,
-        vector=vector,
-        attribution=attribution,
-        metadata={
-            "gate_decision": ctx.gate_decision,
-            "scenario_name": ctx.scenario_name,
-        },
-    )
-
-
-def _hint_feedback_previous_best(ctx: GenerationContext) -> float:
-    """Recover the pre-tournament best score for hint-reflection context."""
-    if ctx.gate_decision == "advance":
-        return max(0.0, ctx.previous_best - ctx.gate_delta)
-    return ctx.previous_best
-
-
-def _collect_hint_feedback(
-    ctx: GenerationContext,
-    *,
-    agents: AgentOrchestrator | None,
-    artifacts: ArtifactStore,
-    sqlite: SQLiteStore,
-    events: EventStreamEmitter,
-) -> HintFeedback | None:
-    """Collect post-tournament competitor feedback on the hints it actually used."""
-    if ctx.settings.ablation_no_feedback or agents is None:
-        return None
-    tournament = ctx.tournament
-    if tournament is None:
-        return None
-    hints_used = ctx.applied_competitor_hints.strip()
-    if not hints_used:
-        return None
-
-    prompt = build_hint_reflection_prompt(
-        hints=hints_used,
-        tournament_best_score=tournament.best_score,
-        tournament_mean_score=tournament.mean_score,
-        previous_best=_hint_feedback_previous_best(ctx),
-    )
-    try:
-        client, resolved_model = agents.resolve_role_execution(
-            "competitor",
-            generation=ctx.generation,
-            scenario_name=ctx.scenario_name,
-        )
-        model = resolved_model or agents.competitor.model
-        response = client.generate(
-            model=model,
-            prompt=prompt,
-            max_tokens=400,
-            temperature=0.2,
-            role="competitor",
-        )
-    except Exception:
-        logger.debug("competitor hint feedback collection failed", exc_info=True)
-        return None
-
-    exec_result = RoleExecution(
-        role="competitor_hint_feedback",
-        content=response.text,
-        usage=RoleUsage(
-            input_tokens=response.usage.input_tokens,
-            output_tokens=response.usage.output_tokens,
-            latency_ms=response.usage.latency_ms,
-            model=response.usage.model,
-        ),
-        subagent_id="competitor_hint_feedback",
-        status="completed",
-    )
-    sqlite.append_generation_agent_activity(
-        ctx.run_id,
-        ctx.generation,
-        outputs=[("competitor_hint_feedback", exec_result.content)],
-        role_metrics=[(
-            exec_result.role,
-            exec_result.usage.model,
-            exec_result.usage.input_tokens,
-            exec_result.usage.output_tokens,
-            exec_result.usage.latency_ms,
-            exec_result.subagent_id,
-            exec_result.status,
-        )],
-    )
-
-    feedback = parse_hint_feedback(response.text, generation=ctx.generation)
-    if feedback.is_empty():
-        return None
-
-    artifacts.write_hint_feedback(ctx.scenario_name, ctx.generation, feedback)
-    events.emit("hint_feedback_collected", {
-        "run_id": ctx.run_id,
-        "generation": ctx.generation,
-        "helpful_count": len(feedback.helpful),
-        "misleading_count": len(feedback.misleading),
-        "missing_count": len(feedback.missing),
-    })
-    return feedback
-
-
-def _hint_volume_policy(ctx: GenerationContext) -> HintVolumePolicy:
-    return HintVolumePolicy(
-        max_hints=ctx.settings.hint_volume_max_hints,
-        archive_rotated=ctx.settings.hint_volume_archive_rotated,
-    )
-
-
-def _hint_feedback_matches(text: str, candidate: str) -> bool:
-    left = text.strip().lower()
-    right = candidate.strip().lower()
-    return bool(left and right and (left == right or left in right or right in left))
-
-
-def _apply_hint_feedback_to_manager(manager: HintManager, feedback: HintFeedback | None) -> None:
-    if feedback is None:
-        return
-    for helpful in feedback.helpful:
-        for hint in manager.active_hints() + manager.archived_hints():
-            if _hint_feedback_matches(hint.text, helpful):
-                manager.update_impact(hint.text, max(hint.impact_score, 0.9))
-    for misleading in feedback.misleading:
-        for hint in manager.active_hints() + manager.archived_hints():
-            if _hint_feedback_matches(hint.text, misleading):
-                manager.update_impact(hint.text, min(hint.impact_score, 0.1))
-
-
-def _update_tool_usage_feedback(
-    ctx: GenerationContext,
-    *,
-    artifacts: ArtifactStore,
-) -> None:
-    """Track competitor tool references so the architect sees real adoption data next gen."""
-    outputs = ctx.outputs
-    if outputs is None or outputs.competitor_output is None:
-        return
-    raw_text = outputs.competitor_output.raw_text
-    if not isinstance(raw_text, str) or not raw_text.strip():
-        return
-
-    known_tools = artifacts.list_tool_names(ctx.scenario_name)
-    if not isinstance(known_tools, list):
-        return
-    tool_names = sorted({name for name in known_tools if isinstance(name, str) and name})
-    if not tool_names:
-        return
-
-    tracker = artifacts.read_tool_usage_tracker(ctx.scenario_name, known_tools=tool_names)
-    if not isinstance(tracker, ToolUsageTracker):
-        tracker = ToolUsageTracker(known_tools=tool_names)
-    tracker.record_generation(ctx.generation, raw_text)
-    artifacts.write_tool_usage_tracker(ctx.scenario_name, tracker)
-
-
-def _maybe_rate_analyst_output(
-    ctx: GenerationContext,
-    *,
-    curator: KnowledgeCurator | None,
-    artifacts: ArtifactStore,
-    sqlite: SQLiteStore,
-) -> AnalystRating | None:
-    """Persist curator feedback on analyst quality when there is a real report to rate."""
-    if curator is None or ctx.settings.ablation_no_feedback:
-        return None
-    outputs = ctx.outputs
-    if outputs is None:
-        return None
-    analysis_markdown = getattr(outputs, "analysis_markdown", "")
-    if not isinstance(analysis_markdown, str) or not analysis_markdown.strip():
-        return None
-
-    tournament = ctx.tournament
-    score_summary = ""
-    if tournament is not None:
-        score_summary = (
-            f"Generation {ctx.generation}: best_score={tournament.best_score:.4f}, "
-            f"mean_score={tournament.mean_score:.4f}, gate_decision={ctx.gate_decision or 'pending'}"
-        )
-    rating, exec_result = curator.rate_analyst_output(
-        analysis_markdown,
-        generation=ctx.generation,
-        score_summary=score_summary,
-        constraint_mode=ctx.settings.constraint_prompts_enabled,
-    )
-    artifacts.write_analyst_rating(ctx.scenario_name, ctx.generation, rating)
-    sqlite.append_generation_agent_activity(
-        ctx.run_id,
-        ctx.generation,
-        outputs=[
-            ("curator_analyst_rating", json.dumps(rating.to_dict(), sort_keys=True)),
-            ("curator_analyst_feedback", exec_result.content),
-        ],
-        role_metrics=[(
-            exec_result.role,
-            exec_result.usage.model,
-            exec_result.usage.input_tokens,
-            exec_result.usage.output_tokens,
-            exec_result.usage.latency_ms,
-            exec_result.subagent_id,
-            exec_result.status,
-        )],
-    )
-    return rating
 
 
 def stage_knowledge_setup(
@@ -2092,7 +934,6 @@ def stage_stagnation_check(
     return ctx
 
 
-
 def stage_skeptic_review(
     ctx: GenerationContext,
     *,
@@ -2264,122 +1105,6 @@ def stage_curator_gate(
     })
 
     return ctx
-
-
-def _persist_skill_note(
-    ctx: GenerationContext,
-    *,
-    artifacts: ArtifactStore,
-) -> None:
-    """Write skill note — advance lessons or rollback warning."""
-    tournament = ctx.tournament
-    assert tournament is not None  # caller guarantees
-    outputs = ctx.outputs
-    assert outputs is not None
-    gate_decision = ctx.gate_decision
-    gate_delta = ctx.gate_delta
-    generation = ctx.generation
-    settings = ctx.settings
-
-    if gate_decision == "advance":
-        skill_lessons = outputs.coach_lessons
-    else:
-        retry_note = f" after {ctx.attempt} retries" if ctx.attempt > 0 else ""
-        skill_lessons = (
-            f"- Generation {generation} ROLLBACK{retry_note} "
-            f"(score={tournament.best_score:.4f}, "
-            f"delta={gate_delta:+.4f}, threshold={settings.backpressure_min_delta}). "
-            f"Strategy: {json.dumps(ctx.current_strategy, sort_keys=True)[:200]}. "
-            f"Narrative: {ctx.replay_narrative[:150]}. "
-            f"Avoid this approach."
-        )
-    artifacts.persist_skill_note(
-        scenario_name=ctx.scenario_name,
-        generation_index=generation,
-        decision=gate_decision,
-        lessons=skill_lessons,
-    )
-
-    # Dead-end registry: record rollback as dead end
-    if gate_decision == "rollback" and settings.dead_end_tracking_enabled:
-        strategy_json = json.dumps(ctx.current_strategy, sort_keys=True)
-        entry = DeadEndEntry.from_rollback(
-            generation=generation,
-            strategy=strategy_json,
-            score=tournament.best_score,
-        )
-        artifacts.append_dead_end(ctx.scenario_name, entry.to_markdown())
-
-
-def _run_curator_consolidation(
-    ctx: GenerationContext,
-    *,
-    curator: KnowledgeCurator,
-    artifacts: ArtifactStore,
-    trajectory_builder: ScoreTrajectoryBuilder,
-    sqlite: SQLiteStore,
-) -> None:
-    """Consolidate lessons and dead-ends via curator."""
-    settings = ctx.settings
-    scenario_name = ctx.scenario_name
-
-    existing_lessons = artifacts.read_skill_lessons_raw(scenario_name)
-    if len(existing_lessons) <= settings.skill_max_lessons:
-        return
-
-    consolidation_trajectory = trajectory_builder.build_trajectory(ctx.run_id)
-    lesson_result, lesson_exec = curator.consolidate_lessons(
-        existing_lessons, settings.skill_max_lessons, consolidation_trajectory,
-        constraint_mode=settings.constraint_prompts_enabled,
-    )
-    artifacts.replace_skill_lessons(scenario_name, lesson_result.consolidated_lessons)
-    sqlite.append_generation_agent_activity(
-        ctx.run_id,
-        ctx.generation,
-        outputs=[("curator_consolidation", lesson_exec.content)],
-        role_metrics=[(
-            lesson_exec.role,
-            lesson_exec.usage.model,
-            lesson_exec.usage.input_tokens,
-            lesson_exec.usage.output_tokens,
-            lesson_exec.usage.latency_ms,
-            lesson_exec.subagent_id,
-            lesson_exec.status,
-        )],
-    )
-
-    # Dead-end consolidation
-    if settings.dead_end_tracking_enabled:
-        dead_end_text = artifacts.read_dead_ends(scenario_name)
-        if dead_end_text:
-            consolidated = consolidate_dead_ends(dead_end_text, max_entries=settings.dead_end_max_entries)
-            artifacts.replace_dead_ends(scenario_name, consolidated)
-
-
-def _persist_progress_snapshot(
-    ctx: GenerationContext,
-    *,
-    artifacts: ArtifactStore,
-) -> None:
-    """Write progress JSON snapshot if enabled."""
-    tournament = ctx.tournament
-    assert tournament is not None  # caller guarantees
-    scenario_name = ctx.scenario_name
-
-    progress_lessons = artifacts.read_skill_lessons_raw(scenario_name)
-    snapshot = build_progress_snapshot(
-        generation=ctx.generation,
-        best_score=ctx.previous_best,
-        best_elo=ctx.challenger_elo,
-        mean_score=tournament.mean_score,
-        gate_history=ctx.gate_decision_history,
-        score_history=ctx.score_history,
-        current_strategy=ctx.current_strategy,
-        lessons=[lesson.lstrip("- ") for lesson in progress_lessons],
-        scoring_backend=tournament.scoring_backend,
-        rating_uncertainty=ctx.challenger_uncertainty,
-    )
-    artifacts.write_progress(scenario_name, snapshot.to_dict())
 
 
 def stage_persistence(
